requirements.txt
altgraph==0.17.5
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.1
pywin32-ctypes==0.2.3
setuptools==82.0.1
